fix: ADL-discoverable int128/double ops; replace long with int64_t
- numeric: move the Boost int128 double-interop shim into namespace
  boost::multiprecision so it is found via ADL. MSVC only consults ADL
  (not ordinary lookup) inside the `requires(Int x, Float g){x*g;}` probe
  in Rational's float constructor, so the global-scope shim was invisible
  there and the constructor still took the infinitely-recursive branch
  (the bentleyottmann stack overflow). ADL placement also lets doctest's
  expression decomposition find int128==double, fixing the C2440 in
  doctest's operator L() when a Boost int128 is a CHECK operand.

- Replace all uses of long / long long with int64_t across the library
  and tests. long is 32-bit on MSVC (LLP64), which overflowed e.g.
  Rational<long> squared distances in the ShapeTree tests. Collapse the
  now-duplicate BigInt operator long/long long into one operator int64_t.

Verified 46/46 on native g++ and clang++ and under a forced Boost-int128
build (g++ -U__SIZEOF_INT128__), including bentleyottmann under a 256 KB
stack (no recursion).

+++ include/algorithm/triangulation.hpp
@@ -1914,8 +1914,8 @@ struct Triangulation {
         std::uint64_t rngState = 0x9e3779b97f4a7c15ULL;
         auto walk = [&](VertexId p, int t) -> int {
             int from = -1;
-            const long cap = 3L * static_cast<long>(tri.size()) + 16;
-            for (long step = 0; step < cap; ++step) {
+            const int64_t cap = int64_t(3) * static_cast<int64_t>(tri.size()) + 16;
+            for (int64_t step = 0; step < cap; ++step) {
                 if (isGhost(t)) {
                     return t;
                 }
